refactor: move and implement the opening of a detailed window
