+++ src/pages/PokemonsPage.module.css
@@ -1,0 +1,11 @@
+.wrapper {
+  display: flex;
+  position: relative;
+}
+
+.hidden {
+  width: 70%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.2);
+  position: absolute;
+}
